GRADLE-2699 - prevent concurrent wrapper downloads from corrupting each other.
This change prevents the zips being downloaded from being corrupted, but does not address race conditions when unpacking the zips.

+++ subprojects/wrapper/src/main/java/org/gradle/wrapper/Install.java
@@ -45,12 +45,18 @@ public class Install {
         File localZipFile = localDistribution.getZipFile();
         boolean downloaded = false;
         if (alwaysDownload || !localZipFile.exists()) {
-            File tmpZipFile = new File(localZipFile.getParentFile(), localZipFile.getName() + ".part");
+            String downloadId = UUID.randomUUID().toString();
+
+            File tmpZipFile = new File(localZipFile.getParentFile(), localZipFile.getName() + "-" + downloadId + ".part");
             tmpZipFile.delete();
             System.out.println("Downloading " + distributionUrl);
             download.download(distributionUrl, tmpZipFile);
-            tmpZipFile.renameTo(localZipFile);
-            downloaded = true;
+
+            if (!localZipFile.exists() || alwaysDownload) {
+                tmpZipFile.renameTo(localZipFile);
+                downloaded = true;
+            }
+
         }
 
         File distDir = localDistribution.getDistributionDir();
